First 2 days
Initially thought of cloning, but I decided to redesign a bit (or a lot?) in the process. First day I worked on the landing page. Did a lot of photoshop/UI design work/research. Tough to use images when not many exists for this company. 2nd page I polished the landing and completed the menu page. Original site feels dark, redundant at times, and sloppy in some areas. I'm aiming to lighten up the design, keep only what is necessary and improve on UX