DONE
Mobile and desktop versions completed. Overall this project turned out to be more of a redesign then clone. I kept the general layout of the site (besides the picture on each page) and made use of color and smart design to clear up, brighten, and make the site easier and more efficient to use